refactor: not required turn class and moved game out of unused game package
