+++ src/main/java/de/markherrmann/javauno/service/GameService.java
@@ -3,7 +3,7 @@ package de.markherrmann.javauno.service;
 import de.markherrmann.javauno.data.fixed.Card;
 import de.markherrmann.javauno.data.fixed.Deck;
 import de.markherrmann.javauno.data.state.UnoState;
-import de.markherrmann.javauno.data.state.component.game.Game;
+import de.markherrmann.javauno.data.state.component.Game;
 import de.markherrmann.javauno.data.state.component.GameLifecycle;
 import de.markherrmann.javauno.data.state.component.Player;
 import org.springframework.beans.factory.annotation.Autowired;
