+++ src/test/java/de/markherrmann/javauno/controller/GameControllerTest.java
@@ -1,7 +1,7 @@
 package de.markherrmann.javauno.controller;
 
 import de.markherrmann.javauno.data.state.UnoState;
-import de.markherrmann.javauno.data.state.component.game.Game;
+import de.markherrmann.javauno.data.state.component.Game;
 import de.markherrmann.javauno.data.state.component.GameLifecycle;
 import de.markherrmann.javauno.data.state.component.Player;
 import org.junit.Test;
