+++ src/test/java/de/markherrmann/javauno/service/GameStateServiceTest.java
@@ -1,7 +1,7 @@
 package de.markherrmann.javauno.service;
 
 import de.markherrmann.javauno.data.state.UnoState;
-import de.markherrmann.javauno.data.state.component.game.Game;
+import de.markherrmann.javauno.data.state.component.Game;
 import de.markherrmann.javauno.data.state.component.GameLifecycle;
 import de.markherrmann.javauno.controller.response.GameState;
 import org.junit.Before;
